+++ src/Symfony/Component/Routing/Matcher/ExpressionLanguageProvider.php
@@ -1,0 +1,41 @@
+<?php
+
+/*
+ * This file is part of the Symfony package.
+ *
+ * (c) Fabien Potencier <fabien@symfony.com>
+ *
+ * For the full copyright and license information, please view the LICENSE
+ * file that was distributed with this source code.
+ */
+
+namespace Symfony\Component\Routing\Matcher;
+
+use Symfony\Component\ExpressionLanguage\ExpressionFunction;
+use Symfony\Component\ExpressionLanguage\ExpressionFunctionProviderInterface;
+
+/**
+ * Defines some ExpressionLanguage functions.
+ *
+ * @author Ahmed TAILOULOUTE <ahmed.tailouloute@gmail.com>
+ */
+class ExpressionLanguageProvider implements ExpressionFunctionProviderInterface
+{
+    /**
+     * {@inheritdoc}
+     */
+    public function getFunctions()
+    {
+        return [
+            new ExpressionFunction(
+                'env',
+                static function ($str, $default = 'null') {
+                    return sprintf('($_SERVER[%s] ?? %s)', $str, $default);
+                },
+                static function ($arguments, $str, $default = null) {
+                    return $_SERVER[$str] ?? $default;
+                }
+            ),
+        ];
+    }
+}